Script count returns int

+++ src/script/Context.cpp
@@ -161,7 +161,7 @@ Context::Context(){
   _globals[Symbol("count")] = (Function)([](const Var&,SymbolVar* stack,size_t params)->Var {
     L_ASSERT(params==1);
     if(stack[0]->is<Ref<Table<Var,Var>>>())
-      return stack[0]->as<Ref<Table<Var,Var>>>()->count();
+      return (int)stack[0]->as<Ref<Table<Var,Var>>>()->count();
     return 0;
   });
   _globals[Symbol("or")] = (Native)([](Context& c,const Array<Var>& a)->Var {
